feat: support media preview in page
1. Add new media preview, you can view images in input panel, instead of click preview icon button;
2. Rename `buildin` to `builtin`.
